requirements.txt
APScheduler==3.7.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lxml @ file:///C:/extend/lxml-4.6.3-cp39-cp39-win_amd64.whl
MarkupSafe==1.1.1
matplotlib @ file:///C:/extend/matplotlib-3.4.1-cp39-cp39-win_amd64.whl
matplotlip==0.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy @ file:///C:/extend/numpy-1.20.2%2Bmkl-cp39-cp39-win_amd64.whl
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow @ file:///C:/extend/Pillow-8.2.0-cp39-cp39-win_amd64.whl
prettytable==2.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.17
pycharts==0.1.5
pycodestyle==2.6.0
pycparser==2.20
pyecharts==1.9.0
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyquery==1.4.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
snownlp==0.12.3
soupsieve==2.2
SQLAlchemy==1.4.7
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wordcloud @ file:///C:/extend/wordcloud-1.8.1-cp39-cp39-win_amd64.whl
WTForms==2.3.3
